update tests for new interface
also note that there's something fishy in libsass. The test.scss that
was provided originally produces different output when compiled by
libsass than when compiled via the ruby compiler

This was already the case with the initial file (@import should all be
moved to the top).

Now as the test.scss as provided didn't really make sense to begin with,
I've reduced it to something that actually compiles correctly.

This is about testing the PHP extension, not libsass